change in user obj in redux
