+++ client/src/redux/store.js
@@ -1,12 +1,12 @@
 import { configureStore } from '@reduxjs/toolkit';
 import auth from 'src/redux/reducers/authSlice';
 import settingsTheme from 'src/redux/reducers/themeSlice';
-// import modal from 'src/redux/reducers/modalSlice';
+import user from 'src/redux/reducers/userSlice';
 
 export const store = configureStore({
   reducer: {
     auth,
     settingsTheme,
-    // modal,
+    user,
   },
 });
